made the filter code better
