+++ code/app/src/main/res/layout/mood_filter.xml
@@ -36,7 +36,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorHappy"
                 android:drawableLeft="@drawable/ic_happy_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="HAPPY"
                 android:textColor="#000000" />
 
@@ -47,7 +46,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorSad"
                 android:drawableLeft="@drawable/ic_sad_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="SAD"
                 android:textColor="#000000" />
 
@@ -58,7 +56,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorSurprised"
                 android:drawableLeft="@drawable/ic_surprised_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="SURPRISED"
                 android:textColor="#000000" />
 
@@ -69,7 +66,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorAfraid"
                 android:drawableLeft="@drawable/ic_afraid_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="AFRAID"
                 android:textColor="#000000" />
 
@@ -80,7 +76,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorDisgusted"
                 android:drawableLeft="@drawable/ic_disgusted_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="DISGUSTED"
                 android:textColor="#000000" />
 
@@ -91,7 +86,6 @@
                 android:layout_gravity="fill"
                 android:backgroundTint="@color/colorAngry"
                 android:drawableLeft="@drawable/ic_angry_icon_24dp"
-                android:onClick="selectMoodButton"
                 android:text="ANGRY"
                 android:textColor="#000000" />
